Add tests to the search service.
We want to add some simple tests around how the search service
behaves so that we have more visibility in the queries being
performed with Elasticsearch.

Since we need to make updates to the query that we're making to accommodate
range queries, it makes sense to start adding some safeguards now.

This introduces (WebMock)[https://github.com/bblimke/webmock/]
and (rspec-snapshots)[https://github.com/yesmeck/rspec-snapshot/]
to provide multiple ways to make smart assertions about the kinds of requests that
Searchkick constructs to send to Elasticsearch under the hood.

We also introduce two new tests to show how this functionality will work.
Once this is merged, we can introduce support for range queries, and
then add tests that will give us a better understanding of how
those queries are constructed from the params the search service receives.